compensation_parameter is small enough for u8

+++ programs/hydra-liquidity-pools/src/lib.rs
@@ -43,7 +43,7 @@ pub mod hydra_liquidity_pools {
         pool_state_bump: u8,
         lp_token_vault_bump: u8,
         lp_token_mint_bump: u8,
-        compensation_parameter: u16,
+        compensation_parameter: u8,
         fees: Fees,
     ) -> Result<()> {
         instructions::initialize::handle(
